Mudança de rota para efetuar a comunicação eficaz entre o FrontEnd e o BackEnd
